+++ core/src/components/checkbox/test/a11y/checkbox.e2e.ts
@@ -12,3 +12,26 @@ configs({ modes: ['ios'], directions: ['ltr'] }).forEach(({ title, config }) => 
     });
   });
 });
+
+configs({ directions: ['ltr'] }).forEach(({ title, config, screenshot }) => {
+  test.describe(title('checkbox: a11y'), () => {
+    test.describe(title('checkbox: font scaling'), () => {
+      test('should scale text on larger font sizes', async ({ page }) => {
+        await page.setContent(
+          `
+            <style>
+              html {
+                font-size: 310%;
+              }
+            </style>
+            <ion-checkbox justify="start" checked>Checked</ion-checkbox>
+          `,
+          config
+        );
+
+        const checkbox = page.locator('ion-checkbox');
+        await expect(checkbox).toHaveScreenshot(screenshot('checkbox-scale'));
+      });
+    });
+  });
+});