+++ core/src/components/range/test/range.spec.ts
@@ -1,5 +1,6 @@
 import { newSpecPage } from '@stencil/core/testing';
 import { Range } from '../range';
+import { Item } from '../../item/item';
 
 let sharedRange;
 describe('Range', () => {
@@ -74,5 +75,154 @@ describe('range id', () => {
 
     const range = page.body.querySelector('ion-range');
     expect(range.getAttribute('id')).toBe('my-custom-range');
+  });
+});
+
+describe('range: item adjustments', () => {
+  it('should add start and end adjustment with no content', async () => {
+    const page = await newSpecPage({
+      components: [Item, Range],
+      html: `
+        <ion-item>
+          <ion-range aria-label="Range"></ion-range>
+        </ion-item>
+      `,
+    });
+
+    const range = page.body.querySelector('ion-range');
+    expect(range.classList.contains('range-item-start-adjustment')).toBe(true);
+    expect(range.classList.contains('range-item-end-adjustment')).toBe(true);
+  });
+
+  it('should add start adjustment with end label and no adornments', async () => {
+    const page = await newSpecPage({
+      components: [Item, Range],
+      html: `
+        <ion-item>
+          <ion-range label="Range" label-placement="end"></ion-range>
+        </ion-item>
+      `,
+    });
+
+    const range = page.body.querySelector('ion-range');
+    expect(range.classList.contains('range-item-start-adjustment')).toBe(true);
+    expect(range.classList.contains('range-item-end-adjustment')).toBe(false);
+  });
+
+  it('should add end adjustment with start label and no adornments', async () => {
+    const page = await newSpecPage({
+      components: [Item, Range],
+      html: `
+        <ion-item>
+          <ion-range label="Range" label-placement="start"></ion-range>
+        </ion-item>
+      `,
+    });
+
+    const range = page.body.querySelector('ion-range');
+    expect(range.classList.contains('range-item-start-adjustment')).toBe(false);
+    expect(range.classList.contains('range-item-end-adjustment')).toBe(true);
+  });
+
+  it('should add end adjustment with fixed label and no adornments', async () => {
+    const page = await newSpecPage({
+      components: [Item, Range],
+      html: `
+        <ion-item>
+          <ion-range label="Range" label-placement="fixed"></ion-range>
+        </ion-item>
+      `,
+    });
+
+    const range = page.body.querySelector('ion-range');
+    expect(range.classList.contains('range-item-start-adjustment')).toBe(false);
+    expect(range.classList.contains('range-item-end-adjustment')).toBe(true);
+  });
+
+  it('should add start adjustment with floating label', async () => {
+    const page = await newSpecPage({
+      components: [Item, Range],
+      html: `
+        <ion-item>
+          <ion-range label="Range" label-placement="floating"></ion-range>
+        </ion-item>
+      `,
+    });
+
+    const range = page.body.querySelector('ion-range');
+    expect(range.classList.contains('range-item-start-adjustment')).toBe(true);
+    expect(range.classList.contains('range-item-end-adjustment')).toBe(true);
+  });
+
+  it('should add start adjustment with stacked label', async () => {
+    const page = await newSpecPage({
+      components: [Item, Range],
+      html: `
+        <ion-item>
+          <ion-range label="Range" label-placement="stacked"></ion-range>
+        </ion-item>
+      `,
+    });
+
+    const range = page.body.querySelector('ion-range');
+    expect(range.classList.contains('range-item-start-adjustment')).toBe(true);
+    expect(range.classList.contains('range-item-end-adjustment')).toBe(true);
+  });
+
+  it('should not add adjustment when not in an item', async () => {
+    const page = await newSpecPage({
+      components: [Item, Range],
+      html: `
+        <ion-range label="Range" label-placement="stacked"></ion-range>
+      `,
+    });
+
+    const range = page.body.querySelector('ion-range');
+    expect(range.classList.contains('range-item-start-adjustment')).toBe(false);
+    expect(range.classList.contains('range-item-end-adjustment')).toBe(false);
+  });
+
+  // TODO FW-2997 remove this
+  it('should not add adjustment with legacy syntax', async () => {
+    const page = await newSpecPage({
+      components: [Item, Range],
+      html: `
+        <ion-range legacy="true"></ion-range>
+      `,
+    });
+
+    const range = page.body.querySelector('ion-range');
+    expect(range.classList.contains('range-item-start-adjustment')).toBe(false);
+    expect(range.classList.contains('range-item-end-adjustment')).toBe(false);
+  });
+
+  it('should not add start adjustment when with start adornment', async () => {
+    const page = await newSpecPage({
+      components: [Item, Range],
+      html: `
+        <ion-range label="Range" label-placement="end">
+          <div slot="start">Start Content</div>
+        </ion-range>
+      `,
+    });
+
+    const range = page.body.querySelector('ion-range');
+    expect(range.classList.contains('range-item-start-adjustment')).toBe(false);
+    expect(range.classList.contains('range-item-end-adjustment')).toBe(false);
+  });
+
+  it('should not add end adjustment when with end adornment', async () => {
+    const page = await newSpecPage({
+      components: [Item, Range],
+      html: `
+        <ion-range label="Range" label-placement="start">
+          <div slot="end">Start Content</div>
+        </ion-range>
+      `,
+    });
+
+    const range = page.body.querySelector('ion-range');
+    expect(range.classList.contains('range-item-start-adjustment')).toBe(false);
+    expect(range.classList.contains('range-item-end-adjustment')).toBe(false);
   });
 });