+++ packages/angular/test/base/src/app/standalone/tabs/tab3.component.ts
@@ -1,0 +1,12 @@
+import { Component } from '@angular/core';
+
+@Component({
+  selector: 'app-tab-three',
+  template: `
+    Tab 3
+  `,
+  standalone: true,
+})
+export class TabThreeComponent {
+
+}